feat(concurrency): Refactor test suites
