+++ dd-java-agent/instrumentation/java-concurrent/src/test/groovy/executor/RejectedExecutionTest.groovy
@@ -1,8 +1,13 @@
+package executor
+
 import datadog.trace.agent.test.AgentTestRunner
 import datadog.trace.api.DDSpanId
 import datadog.trace.core.DDSpan
+import executor.rejectedexecutionhandler.ExecutingRejectedExecutionHandler
+import executor.rejectedexecutionhandler.SwallowingRejectedExecutionHandler
 import io.netty.util.concurrent.DefaultEventExecutor
 import io.netty.util.concurrent.DefaultThreadFactory
+import runnable.JavaAsyncChild
 
 import java.util.concurrent.ArrayBlockingQueue
 import java.util.concurrent.CountDownLatch
