+++ dd-java-agent/instrumentation/java-concurrent/src/test/groovy/forkjoin/ForkJoinPoolPropagationTest.groovy
@@ -1,3 +1,5 @@
+package forkjoin
+
 import datadog.trace.agent.test.AgentTestRunner
 import datadog.trace.core.DDSpan
