+++ dd-java-agent/instrumentation/java-concurrent/src/test/groovy/runnable/ModuleInjectionTest.groovy
@@ -1,3 +1,5 @@
+package runnable
+
 import datadog.trace.agent.test.AgentTestRunner
 
 import javax.swing.*
