+++ dd-java-agent/instrumentation/java-concurrent/src/test/groovy/timer/TimerTaskContinuationTest.groovy
@@ -1,3 +1,5 @@
+package timer
+
 import static datadog.trace.agent.test.utils.TraceUtils.basicSpan
 import static datadog.trace.agent.test.utils.TraceUtils.runUnderTrace
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activateSpan
