+++ dd-java-agent/instrumentation/java-concurrent/src/test/java/executor/CustomThreadPoolExecutor.java
@@ -1,3 +1,5 @@
+package executor;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.AbstractExecutorService;
