+++ dd-java-agent/instrumentation/java-concurrent/src/test/java/executor/TypeAwareThreadPoolExecutor.java
@@ -1,3 +1,5 @@
+package executor;
+
 import datadog.trace.bootstrap.instrumentation.java.concurrent.Wrapper;
 import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.ThreadPoolExecutor;
