+++ dd-java-agent/instrumentation/java-concurrent/src/test/java/executor/recursive/RecursiveThreadPoolMixedSubmissionAndExecution.java
@@ -1,3 +1,5 @@
+package executor.recursive;
+
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activateSpan;
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.startSpan;
 
@@ -5,14 +7,15 @@ import datadog.trace.bootstrap.instrumentation.api.AgentScope;
 import datadog.trace.bootstrap.instrumentation.api.AgentSpan;
 import java.util.concurrent.ExecutorService;
 
-public class RecursiveThreadPoolSubmission implements Runnable {
+public class RecursiveThreadPoolMixedSubmissionAndExecution implements Runnable {
 
   private final ExecutorService executor;
 
   private final int maxDepth;
   private final int depth;
 
-  public RecursiveThreadPoolSubmission(ExecutorService executor, int maxDepth, int depth) {
+  public RecursiveThreadPoolMixedSubmissionAndExecution(
+      ExecutorService executor, int maxDepth, int depth) {
     this.executor = executor;
     this.maxDepth = maxDepth;
     this.depth = depth;
@@ -26,7 +29,13 @@ public class RecursiveThreadPoolSubmission implements Runnable {
     AgentSpan span = startSpan(String.valueOf(depth));
     try (AgentScope scope = activateSpan(span)) {
       scope.setAsyncPropagation(true);
-      executor.submit(new RecursiveThreadPoolSubmission(executor, maxDepth, depth + 1));
+      if (depth % 2 == 0) {
+        executor.submit(
+            new RecursiveThreadPoolMixedSubmissionAndExecution(executor, maxDepth, depth + 1));
+      } else {
+        executor.execute(
+            new RecursiveThreadPoolMixedSubmissionAndExecution(executor, maxDepth, depth + 1));
+      }
     } finally {
       span.finish();
     }
