+++ dd-java-agent/instrumentation/java-concurrent/src/test/java/executor/rejectedexecutionhandler/ExecutingRejectedExecutionHandler.java
@@ -1,3 +1,5 @@
+package executor.rejectedexecutionhandler;
+
 import java.util.concurrent.RejectedExecutionHandler;
 import java.util.concurrent.ThreadPoolExecutor;
