+++ dd-java-agent/instrumentation/java-concurrent/src/test/java/executor/rejectedexecutionhandler/SwallowingRejectedExecutionHandler.java
@@ -1,3 +1,5 @@
+package executor.rejectedexecutionhandler;
+
 import io.netty.util.concurrent.SingleThreadEventExecutor;
 import java.util.concurrent.RejectedExecutionHandler;
 import java.util.concurrent.ThreadPoolExecutor;
