+++ dd-java-agent/instrumentation/java-concurrent/src/test/java/forkjoin/LinearTask.java
@@ -1,3 +1,5 @@
+package forkjoin;
+
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activateSpan;
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.startSpan;
