+++ dd-java-agent/instrumentation/java-concurrent/src/test/java/forkjoin/PeriodicTask.java
@@ -1,3 +1,5 @@
+package forkjoin;
+
 import datadog.trace.api.Trace;
 
 public class PeriodicTask implements Runnable {
