+++ dd-java-agent/instrumentation/java-concurrent/src/test/java/runnable/ComparableAsyncChild.java
@@ -1,3 +1,5 @@
+package runnable;
+
 public class ComparableAsyncChild implements Runnable, Comparable<ComparableAsyncChild> {
 
   private final int priority;
