+++ dd-java-agent/instrumentation/java-concurrent/src/test/java/runnable/Descendant.java
@@ -1,3 +1,5 @@
+package runnable;
+
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.activateSpan;
 import static datadog.trace.bootstrap.instrumentation.api.AgentTracer.startSpan;
