+++ dd-java-agent/instrumentation/java-concurrent/src/test/java/runnable/JavaAsyncChild.java
@@ -1,3 +1,5 @@
+package runnable;
+
 import datadog.trace.api.Trace;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ForkJoinTask;
