+++ dd-java-agent/instrumentation/java-concurrent/src/test/java/runnable/SettableFuture.java
@@ -1,3 +1,5 @@
+package runnable;
+
 import java.util.concurrent.Callable;
 import java.util.concurrent.FutureTask;
